# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sentinel RMM API
  description: |
    Remote Monitoring and Management API for Sentinel RMM.

    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Tokens are obtained via the `/api/auth/login` endpoint.

    ## Rate Limiting
    API requests are rate limited. Default limits are 100 requests per minute.
  version: 1.0.0
  contact:
    name: Sentinel RMM Support
  license:
    name: Proprietary
servers:
  - url: /api
    description: API Base Path
security:
  - bearerAuth: []
...
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: List all devices
      description: Retrieve a list of all managed devices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
...
  /scripts:
    get:
      tags: [Scripts]
      summary: List scripts
      responses:
        '200':
          description: Script list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Script'
    post:
      tags: [Scripts]
      summary: Create script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptCreate'
      responses:
        '201':
          description: Script created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostname:
          type: string
        osType:
          type: string
        osVersion:
          type: string
        status:
          type: string
          enum: [online, offline]
        ipAddress:
          type: string
        lastSeen:
          type: string
          format: date-time
        agentVersion:
          type: string
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        severity:
          type: string
          enum: [critical, warning, info]
        title:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, acknowledged, resolved]
        createdAt:
          type: string
          format: date-time
...
    Script:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        language:
          type: string
          enum: [powershell, bash, python]
        content:
          type: string
        isPreset:
          type: boolean
    ScriptCreate:
      type: object
      required: [name, language, content]
      properties:
        name:
          type: string
        description:
          type: string
        language:
          type: string
          enum: [powershell, bash, python]
        content:
          type: string
...